Add customization for Quick Settings [1/2]
Allow Quick Settings tiles to be arranged any way you want, delete them or add some.

I removed the "dual" status that was hardcoded into Wifi and BT tiles and added it
as a variable which is true when the tile is on the first row.

Also, the first row is now hardcoded to have only 2 items to follow the same design as before.

Inversion and Hotspot tile are no longer contextual and can be added, rmeoved or rearranged at will.

Settings: http://review.cyanogenmod.org/83418

Change-Id: Ie596a54a5c4eb777782fb047496cfcd0a6cd2e89

SystemUI: Update SystemUI to point to new QS secure setting.

Change-Id: Ieb11b257d9fbf6a9fc046d76a61eafa307c1df72

Conflicts:
	packages/SystemUI/src/com/android/systemui/qs/QSPanel.java

Migrate QS_TILES to Settings.Secure

Change-Id: I42783f574d7e050a04e960c91883f2aefc8ab7d6

[Crostantin]: adapted for BoostPop

Signed-off-by: Crostantin <crostantin@gmail.com>